requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
openai>=1.30.0
google-genai==1.33.0
python-multipart

deepface==0.0.93
tf-keras
opencv-python-headless==4.10.0.84
numpy==1.26.4

